refactor : 사용 안하는 주석, 메서드 제거
