--- src/main/java/caffeine/nest_dev/common/websocket/dto/SessionInfo.java
@@ -1,13 +1,0 @@
-package caffeine.nest_dev.common.websocket.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class SessionInfo {
-
-    private String sessionId;
-}
