--- src/main/java/caffeine/nest_dev/common/websocket/util/StompPrincipal.java
@@ -1,18 +1,0 @@
-package caffeine.nest_dev.common.websocket.util;
-
-
-import java.security.Principal;
-
-public class StompPrincipal implements Principal {
-
-    private final String name;
-
-    public StompPrincipal(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String getName() {
-        return name;
-    }
-}
