+++ src/main/java/caffeine/nest_dev/domain/chatroom/scheduler/event/SaveTerminationRoomEvent.java
@@ -1,4 +1,4 @@
-package caffeine.nest_dev.domain.chatroom.scheduler.util;
+package caffeine.nest_dev.domain.chatroom.scheduler.event;
 
 import caffeine.nest_dev.domain.reservation.entity.Reservation;
 import java.time.LocalDateTime;
@@ -10,15 +10,15 @@ import lombok.Getter;
 @Getter
 @Builder
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class SaveCreateRoomEvent {
+public class SaveTerminationRoomEvent {
 
-    private final Long reservationId;
-    private final LocalDateTime startAt;
+    private final long reservationId;
+    private final LocalDateTime endAt;
 
-    public static SaveCreateRoomEvent from(Reservation reservation) {
-        return SaveCreateRoomEvent.builder()
+    public static SaveTerminationRoomEvent from(Reservation reservation) {
+        return SaveTerminationRoomEvent.builder()
                 .reservationId(reservation.getId())
-                .startAt(reservation.getReservationStartAt())
+                .endAt(reservation.getReservationEndAt())
                 .build();
     }
 }
