+++ src/main/java/caffeine/nest_dev/domain/notification/dto/response/NotificationResponseDto.java
@@ -2,13 +2,14 @@ package caffeine.nest_dev.domain.notification.dto.response;
 
 import caffeine.nest_dev.domain.notification.entity.Notification;
 import java.time.LocalDateTime;
+import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @Builder
-@AllArgsConstructor
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class NotificationResponseDto {
 
     private Long id;
